+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-Version 0.3.9
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -11,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - A new token ref at pda "token-ref", owner as the primary token ref
   - SetAsPrimaryV0 endpoint and related sdk updates
   - `curveAuthority` and `reserveAuthority` on token bonding instances that will in the future allow changes to the curve and changes to the reserves account, respectively.
+  - Ability to `updateCollective` to change both the config and authority.
 
 ### Changed
   -